+++ app/models/taxon_name_classification/latinized/gender.rb
@@ -2,6 +2,8 @@ class TaxonNameClassification::Latinized::Gender < TaxonNameClassification::Lati
 
   NOMEN_URI='http://purl.obolibrary.org/obo/NOMEN_0000045'.freeze
 
+  validate :validate_applicable_rank
+
   def self.applicable_ranks
     GENUS_RANK_NAMES
   end
@@ -13,6 +15,14 @@ class TaxonNameClassification::Latinized::Gender < TaxonNameClassification::Lati
   end
 
   private
+
+  def validate_applicable_rank
+    return true if taxon_name.nil?
+
+    unless self.class.applicable_ranks.include?(taxon_name.rank_string)
+      errors.add(:taxon_name, 'Gender is only applicable to genus names')
+    end
+  end
 
   def set_cached_names_for_taxon_names
     t = taxon_name